# sift: optimization stage1