**Tested with main PR**: https://github.com/opencv/opencv/pull/17092

- perf tests
- trace regions
- avoid inplace arrays in `GaussianBlur()` calls

<details>

<summary>SIFT performance results (i5-6600)</summary>

|Name of Test|base|ipp|avx2+ipp|ipp speedup|ipp+avx2 speedup|
|---|:-:|:-:|:-:|:-:|:-:|
|SIFT::SIFT_detect::"leuven/img1.png"|79.523|34.281|34.230|2.32|2.32|
|SIFT::SIFT_detect::"stitching/a3.png"|71.424|31.850|31.629|2.24|2.26|
|SIFT::SIFT_extract::"leuven/img1.png"|78.763|34.135|26.981|2.31|2.92|
|SIFT::SIFT_extract::"stitching/a3.png"|85.059|46.014|33.927|1.85|2.51|
|SIFT::SIFT_full::"leuven/img1.png"|100.167|55.374|48.046|1.81|2.08|
|SIFT::SIFT_full::"stitching/a3.png"|106.117|66.773|54.400|1.59|1.95|
</details>

<!--
allow_multiple_commits=1
-->